pc_buyer_review_new/requirements.txt
altair==5.5.0
amazon-textract-caller==0.2.4
amazon-textract-response-parser==1.0.3
amazon-textract-textractor==1.9.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
boto3==1.37.38
botocore==1.37.38
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.0
editdistance==0.8.1
Events==0.5
exceptiongroup==1.2.2
faiss-cpu==1.11.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
inflect==7.5.0
Jinja2==3.1.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.23
langchain-core==0.3.55
langchain-text-splitters==0.3.8
langsmith==0.3.33
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.6.0
narwhals==1.39.0
numpy==2.2.5
opensearch-py==2.8.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==6.31.0
pyarrow==20.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-aws4auth==1.3.1
requests-toolbelt==1.0.0
rpds-py==0.24.0
s3transfer==0.11.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
streamlit==1.45.1
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typeguard==4.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
XlsxWriter==3.2.3
zstandard==0.23.0
